requirements.txt
pandas
numpy
matplotlib
seaborn
yfinance
pandas-datareader
scikit-learn
keras
tensorflow
dvc
notebook
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.4
joblib
types-pyYAML
Flask 
Flask-Cors
scipy
psx-data-reader
-e .